backend/requirements.txt
django>=5.2
djangorestframework>=3.16
djangorestframework-simplejwt>=5.3
